+++ frontend/src/__tests__/components/common/Form/Form.test.js
@@ -1,5 +1,5 @@
 import { render,screen } from '@testing-library/react';
-import Form from '../../../../components/common/Form/Form';
+import Form from '../../../../components/Form/Form';
 
 test('renders the Form component correctly', () => {
     const formProps = {
