+++ frontend/src/components/Form/Form.tsx
@@ -6,7 +6,7 @@ const Form = (props: any) => {
     for (const field in props.fields) {
         const inputField = props.fields[field];
         inputFields.push(
-            <input placeholder={field} {...inputField} key={field}/>
+            <input placeholder={field} {...inputField} key={field} />
         );
     }
 
@@ -14,7 +14,7 @@ const Form = (props: any) => {
         <div className='form'>
             <form>
                 {inputFields}
-                <input type='submit' className='submit-button' value='Submit' onClick={props.onSubmit}/>
+                <input type='submit' className='submit-button' value='Submit' onClick={props.onSubmit} />
             </form>
         </div>
     );
