+++ frontend/src/components/LogIn/LogIn.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
-import Form from "../common/Form/Form";
+import { useState } from 'react';
+import Form from '../Form/Form';
 import authenticate from '../../utils/authenticate';
 
 export default function LogIn(props: any) {
